@@ -20,7 +20,7 @@ zip_safe = False
 packages = find_namespace:
 package_dir =
   = src
-python_requires = >= 3.6
+python_requires = >= 3.7
 install_requires =
   param>=1.9
 
