@@ -1,5 +1,6 @@
 from shutil import rmtree
 from tempfile import mkdtemp
+from io import StringIO
 
 import pytest
 import param
@@ -12,17 +13,20 @@ param.parameterized.warnings_as_exceptions = True
 @pytest.fixture(params=["ruamel_yaml", "pyyaml"])
 def yaml_loader(request):
     if request.param == "ruamel_yaml":
+        YAML = None
         try:
             from ruamel_yaml import YAML  # type: ignore
-
-            yaml_loader = YAML().load
         except ImportError:
-            from ruamel.yaml import YAML  # type: ignore
-
-            yaml_loader = YAML().load
+            pass
+        if YAML is None:
+            try:
+                from ruamel.yaml import YAML  # type: ignore
+            except ImportError:
+                pytest.skip("No yaml parser found")
+        yaml_loader = YAML().load
         module_names = ("ruamel_yaml", "ruamel.yaml")
     else:
-        import yaml  # type: ignore
+        yaml = pytest.importorskip("yaml")
 
         def yaml_loader(x):
             return yaml.load(x, Loader=yaml.FullLoader)
@@ -37,6 +41,11 @@ def yaml_loader(request):
 @pytest.fixture(params=[True, False])
 def with_yaml(request):
     if request.param:
+        try:
+            with StringIO() as fp:
+                serial._serialize_to_yaml(fp, {"foo": 1})
+        except ImportError:
+            pytest.skip("No yaml serializer")
         yield True
     else:
         old_props = serial.YAML_MODULE_PRIORITIES
